requirements.txt
pandas
numpy
scikit-learn
xgboost
streamlit
plotly
joblib
kaggle
openmeteo-requests
requests-cache
retry-requests
statsmodels